Split fixtures in channel module
