+++ saleor/channel/tests/fixtures/__init__.py
@@ -1,0 +1,1 @@
+from .channel import *  # noqa: F403
