+++ saleor/graphql/channel/tests/benchmark/__init__.py
@@ -1,0 +1,1 @@
+CHANNEL_COUNT_IN_BENCHMARKS = 10
